deprecated_files/old_requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bert-score==0.3.13
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.19.0
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
evaluate==0.4.2
exceptiongroup==1.2.1
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
googleapis-common-protos==1.63.0
googletrans==3.0.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.22.2
hyperframe==5.2.0
idna==2.10
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.50
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.23.3
openpyxl==3.1.2
orjson==3.10.1
packaging==23.2
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
Pygments==2.17.2
pymongo==4.7.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
streamlit==1.33.0
streamlit-option-menu==0.3.12
sympy==1.12
tenacity==8.2.3
textaugment==2.0.0
textblob==0.18.0.post0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
transformers==4.40.1
Trial2Vec==0.1.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wget==3.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
